backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["config*", "core*", "accounts*", "notes*"]

[project]
name = "turboai-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.5",
    "django-cors-headers>=4.7",
    "drf-pydantic>=2.9",
    "psycopg[binary,pool]>=3.1.8",
    "django-redis>=6.0",
    "uvicorn[standard]",
    "django-environ",
    "nh3>=0.3.3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-django",
]

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["config", "core", "accounts", "notes"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
